@@ -1,6 +1,6 @@
 name: flutter_openim_widget
 description: The public ui library is used with the openim demo, and you can directly use it for secondary development.
-version: 2.1.0
+version: 2.3.0
 homepage: https://github.com/hrxiang/flutter_openim_widget.git
 
 environment:
@@ -37,9 +37,14 @@ dependencies:
   mime_type: ^1.0.0
   dio: ^4.0.4 #网络s请求
   flutter_keyboard_visibility: ^5.3.0
-  flutter_openim_sdk: ^2.2.0
+  emoji_picker_flutter: ^1.3.0
+  should_rebuild: ^1.0.1
+  flutter_openim_sdk: ^2.3.0
 #  flutter_openim_sdk:
 #    path: ../flutter_openim_sdk
+#    git:
+#      url: https://github.com/OpenIMSDK/Open-IM-SDK-Flutter.git
+
 
 
 dev_dependencies:
