@@ -39,9 +39,9 @@ export 'src/chat_voice_record_view.dart';
 export 'src/chat_voice_view.dart';
 export 'src/chat_webview_map.dart';
 export 'src/conversation_itemview.dart';
-export 'src/copy_custom_pop_up_menu.dart';
-export 'src/copy_focus_detector.dart';
 export 'src/custom_chawie_controls.dart';
+export 'src/custom_focus_detector.dart';
+export 'src/custom_pop_up_menu.dart';
 export 'src/favorite_emoji_listview.dart';
 export 'src/pop_button.dart';
 export 'src/timing_view.dart';
