@@ -1,65 +1,61 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class UnreadCountView extends StatelessWidget {
-  final Color color;
-  final double size;
-  final Stream<int>? steam;
-  final int? count;
-
   const UnreadCountView({
     Key? key,
-    this.steam,
-    this.color = const Color(0xFFF44038),
-    this.size = 13,
     this.count = 0,
+    this.size = 13,
+    this.margin,
   }) : super(key: key);
+  final int count;
+  final double size;
+  final EdgeInsetsGeometry? margin;
 
   @override
   Widget build(BuildContext context) {
-    if (null == steam) {
-      return _buildChild(count: count ?? 0);
-    }
-    return StreamBuilder(
-      stream: steam,
-      builder: (_, AsyncSnapshot<int> hot) => Visibility(
-        visible: (hot.data ?? 0) > 0,
-        child: Container(
-          width: size,
-          height: size,
-          alignment: Alignment.center,
-          decoration: BoxDecoration(
-            color: color,
-            shape: BoxShape.circle,
-          ),
-          child: Text(
-            '${(hot.data ?? 0) > 99 ? '...' : hot.data}',
-            style: TextStyle(
-              fontSize: 8,
-              color: Color(0xFFFFFFFF),
-            ),
-          ),
-        ),
+    return Visibility(
+      visible: count > 0,
+      child: Container(
+        alignment: Alignment.center,
+        margin: margin,
+        padding: count > 99 ? EdgeInsets.symmetric(horizontal: 4.w) : null,
+        constraints: BoxConstraints(maxHeight: size, minWidth: size),
+        decoration: _decoration,
+        child: _text,
       ),
     );
   }
 
-  Widget _buildChild({required int count}) => Visibility(
-    visible: count > 0,
-    child: Container(
-      width: size,
-      height: size,
-      alignment: Alignment.center,
-      decoration: BoxDecoration(
-        color: color,
-        shape: BoxShape.circle,
-      ),
-      child: Text(
-        '${count > 99 ? '...' : count}',
+  Text get _text => Text(
+        '${count > 99 ? '99+' : count}',
         style: TextStyle(
-          fontSize: 8,
-          color: Color(0xFFFFFFFF),
+          fontSize: 8.sp,
+          color: const Color(0xFFFFFFFF),
         ),
-      ),
-    ),
-  );
+        textAlign: TextAlign.center,
+      );
+
+  Decoration get _decoration => BoxDecoration(
+        color: Color(0xFFFF3366),
+        shape: count > 99 ? BoxShape.rectangle : BoxShape.circle,
+        borderRadius: count > 99 ? BorderRadius.circular(10.r) : null,
+        boxShadow: [
+          BoxShadow(
+            color: Color(0x26C61B4A),
+            offset: Offset(1.15.w, 1.15.h),
+            blurRadius: 57.58.r,
+          ),
+          BoxShadow(
+            color: Color(0x1AC61B4A),
+            offset: Offset(2.3.w, 2.3.h),
+            blurRadius: 11.52.r,
+          ),
+          BoxShadow(
+            color: Color(0x0DC61B4A),
+            offset: Offset(4.61.w, 4.61.h),
+            blurRadius: 17.28.r,
+          ),
+        ],
+      );
 }
